put functions only used in testing in a new file
