+++ src/neo4j/neo4j-functions.js
@@ -1,9 +1,6 @@
 import neo4j from 'neo4j-driver';
-import {
-    giveConnectedInfoByGeneId, makeNodeQuery, makeRelationshipQuery,
-    numNodes, numEdges, deleteAll, returnEdgeArticleTitleById
-} from './query-strings';
-import { closeDriver, getDriver, initDriver } from './neo4j-driver';
+import { giveConnectedInfoByGeneId, makeNodeQuery, makeRelationshipQuery } from './query-strings';
+import { getDriver } from './neo4j-driver';
 
 export async function addNode(id, name) {
     const driver = getDriver();
@@ -76,82 +73,5 @@ export async function searchByGeneId(id) {
         await session.close();
     }
     return;
-}
-
-export async function getEdgeArticleTitleById(id) {
-    const driver = getDriver();
-    let session;
-    let articleTitle;
-    try {
-        session = driver.session({ database: 'neo4j' });
-        let result = session.executeRead(tx => {
-            return tx.run(returnEdgeArticleTitleById, { id: id });
-        });
-        articleTitle = result.records.get('articleTitle');
-    } catch (error) {
-        console.error(error);
-        throw error;
-    } finally {
-        await session.close();
-    }
-    return articleTitle;
-}
-
-export async function deleteAllNodesAndEdges() {
-    const driver = getDriver();
-    let session;
-    try {
-        session = driver.session({ database: "neo4j" });
-        let result = await session.executeWrite(tx => {
-            return tx.run(deleteAll);
-        });
-    } catch (error) {
-        console.error(error);
-        throw error;
-    } finally {
-        await session.close();
-    }
-    return;
-}
-
-export async function getNumNodes() {
-    const driver = getDriver();
-    let session;
-    let num;
-    try {
-        session = driver.session({ database: "neo4j" });
-        let result = await session.executeRead(tx => {
-            return tx.run(numNodes);
-        });
-        num = result.records[0].get(0).toNumber();
-        console.log(num.toString());
-    } catch (error) {
-        console.error(error);
-        throw error;
-    } finally {
-        await session.close();
-    }
-    return num;
-}
-
-export async function getNumEdges() {
-    const driver = getDriver();
-    let session;
-    let num;
-    try {
-        session = driver.session({ database: "neo4j" });
-        let result = await session.executeRead(tx => {
-            return tx.run(numEdges);
-        });
-        num = result.records.map(row => {
-            return row.get('count');
-        });
-    } catch (error) {
-        console.error(error);
-        throw error;
-    } finally {
-        await session.close();
-    }
-    return num;
 }
